finish member_manage

+++ association_member_manager/src/main/java/exort/association_member_management/entity/DepartmentId.java
@@ -11,5 +11,5 @@ public class DepartmentId implements Serializable {
     @NotNull
     int associationId;
     @NotNull
-    long departmentId;
+    int departmentId;
 }
